feat(deploy): add Grafana to docker-compose for monitoring
Add Grafana to the docker-compose configuration to enable monitoring
of the APISIX dashboard. This commit includes the setup of Grafana
with a volume for data persistence, environment variables for
authentication and security settings, and mapping of the Grafanaport to host's port 3000. Grafana is set to always restart and join
the 'z-chat' network.